Improve Dockerfile
Added environment variables so that the Dogecoin Core version, SHA256
checksum and download URL could be more easily changed in the future.

Removed steps which fetched unnecessary packages for build/runtime.

In rosetta-builder, separate fetching dependencies and building the go
binary steps so that they can be cached if the dependencies haven't
changed.